fix: changed mainPageContainer
