+++ reactjs-films-homework/src/components/not-found-component/NotFoundComponent.jsx
@@ -1,0 +1,7 @@
+import styles from "./not-found-style.module.scss";
+
+const NotFoundComponent = ({ message }) => {
+  return <h2 className={styles.text}>{message}</h2>;
+};
+
+export default NotFoundComponent;
