+++ reactjs-films-homework/src/components/not-found-component/not-found-style.module.scss
@@ -1,0 +1,10 @@
+.text {
+  display: block;
+  position: absolute;
+  top: 50%;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  margin: auto;
+  text-align: center;
+}
