+++ reactjs-films-homework/src/pages/main-page/MainPageContainer.jsx
@@ -1,34 +1,14 @@
 import MainPage from "./MainPage";
-import { useDispatch, useSelector } from "react-redux";
-import { useEffect } from "react";
-import {
-  fetchMovies,
-  fetchSearchMovies,
-} from "../../store/asyncThunks/mainPage/asunkThunksMainPage";
 import DownloadSpinner from "../../components/download-spinner/DownloadSpinner";
+import NotFoundComponent from "../../components/not-found-component/NotFoundComponent";
+
+import { SideEffect } from "./mainPageSideEffect";
 
 const MainPageContainer = ({ setPage }) => {
-  const dispatch = useDispatch();
-
-  const {
-    status,
-    topMovies,
-    language,
-    search,
-    searchMovies,
-    api_category,
-    category,
-    currentPage,
-  } = useSelector((state) => state.mainPageSlice);
-
-  useEffect(() => {
-    if (search.length === 0)
-      dispatch(fetchMovies({ language, api_category, currentPage }));
-
-    if (search.length > 0) dispatch(fetchSearchMovies({ language, search }));
-  }, [dispatch, language, search, category, api_category, currentPage]);
-
-  if (status === "loading" || status === "loading") return <DownloadSpinner />;
+  const funcSideEffect = SideEffect();
+  const { status, topMovies, search, searchMovies, error } = funcSideEffect;
+  if (status === "loading") return <DownloadSpinner />;
+  if (status === "rejected") return <NotFoundComponent message={error} />;
   return (
     <MainPage
       setPage={setPage}
