+++ reactjs-films-homework/src/pages/main-page/mainPageSideEffect.js
@@ -1,0 +1,34 @@
+import { useDispatch, useSelector } from "react-redux";
+import { useEffect } from "react";
+import {
+  fetchMovies,
+  fetchSearchMovies,
+} from "../../store/asyncThunks/mainPage/asunkThunksMainPage";
+
+export function SideEffect() {
+  const dispatch = useDispatch();
+  const {
+    status,
+    topMovies,
+    language,
+    search,
+    searchMovies,
+    api_category,
+    category,
+    currentPage,
+    error,
+  } = useSelector((state) => state.mainPageSlice);
+  useEffect(() => {
+    if (search.length === 0)
+      dispatch(fetchMovies({ language, api_category, currentPage }));
+
+    if (search.length > 0) dispatch(fetchSearchMovies({ language, search }));
+  }, [dispatch, language, search, category, api_category, currentPage]);
+  return {
+    status,
+    topMovies,
+    search,
+    searchMovies,
+    error,
+  };
+}
